backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
av==15.1.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
elevenlabs==2.24.0
eval_type_backport==0.3.0
executing==2.2.1
fastapi==0.121.3
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jupyter_client==8.6.3
jupyter_core==5.9.1
livekit==1.0.19
livekit-agents==1.3.3
livekit-api==1.0.7
livekit-blingfire==1.0.0
livekit-plugins-deepgram==1.3.3
livekit-plugins-elevenlabs==1.3.3
livekit-plugins-openai==1.3.3
livekit-plugins-silero==1.3.3
livekit-protocol==1.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.1
openai==2.8.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
silero-vad==6.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
torch==2.9.1
torchaudio==2.9.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
types-protobuf==6.32.1.20251105
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
